+++ configs/eslint-config-js/index.js
@@ -1,12 +1,12 @@
 /** @type {import('eslint').Linter.Config} */
 module.exports = {
   root: true,
-  extends: ['plugin:import/recommended', './no-import.js'],
+  extends: ['plugin:import/recommended', 'plugin:import/recommended', './no-import.js'],
   plugins: ['import'],
   rules: {
     'sort-imports': ['error', { ignoreDeclarationSort: true }],
-    'import/no-duplicates': 'error',
     'import/no-cycle': 'error',
+    'import/no-duplicates': ['error', { 'prefer-inline': true }],
     'import/order': [
       'warn',
       {
