+++ configs/eslint-config-ts/index.js
@@ -1,7 +1,7 @@
 /** @type {import('eslint').Linter.Config} */
 module.exports = {
   root: true,
-  extends: ['plugin:import/typescript', './no-import.js'],
+  extends: ['plugin:import/recommended', 'plugin:import/typescript', './no-import.js'],
   plugins: ['import'],
   settings: {
     'import/resolver': {
@@ -16,8 +16,8 @@ module.exports = {
   ],
   rules: {
     'sort-imports': ['error', { ignoreDeclarationSort: true }],
-    'import/no-duplicates': 'error',
     'import/no-cycle': 'error',
+    'import/no-duplicates': ['error', { 'prefer-inline': true }],
     'import/order': [
       'error',
       {
