+++ packages/react-image/src/Image.tsx
@@ -1,5 +1,4 @@
-import type { ComponentProps } from 'react'
-import { forwardRef } from 'react'
+import { type ComponentProps, forwardRef } from 'react'
 import { useLoad } from './Load'
 
 type ImageProps = ComponentProps<'img'>
