+++ packages/react/src/AsyncBoundary.spec.tsx
@@ -1,8 +1,7 @@
 import { ERROR_MESSAGE, FALLBACK, Suspend, TEXT, ThrowError } from '@suspensive/test-utils'
 import { act, render, waitFor } from '@testing-library/react'
 import ms from 'ms'
-import type { ComponentProps } from 'react'
-import { createElement } from 'react'
+import { type ComponentProps, createElement } from 'react'
 import { createRoot } from 'react-dom/client'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 import { AsyncBoundary, withAsyncBoundary } from '.'
