+++ packages/react/src/ErrorBoundary.spec.tsx
@@ -1,8 +1,7 @@
 import { ERROR_MESSAGE, FALLBACK, TEXT, ThrowError, ThrowNull } from '@suspensive/test-utils'
 import { act, render } from '@testing-library/react'
 import ms from 'ms'
-import type { ComponentProps, ComponentRef } from 'react'
-import { createElement, createRef } from 'react'
+import { type ComponentProps, type ComponentRef, createElement, createRef } from 'react'
 import { createRoot } from 'react-dom/client'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 import { useTimeout } from './hooks'
