+++ packages/react/src/ErrorBoundaryGroup.tsx
@@ -1,5 +1,12 @@
-import type { PropsWithChildren, ReactNode } from 'react'
-import { createContext, useContext, useEffect, useMemo, useReducer } from 'react'
+import {
+  type PropsWithChildren,
+  type ReactNode,
+  createContext,
+  useContext,
+  useEffect,
+  useMemo,
+  useReducer,
+} from 'react'
 import { useIsChanged } from './hooks'
 import { assert, increase } from './utils'
