+++ packages/react/src/contexts/index.ts
@@ -1,1 +1,2 @@
 export { DelayDefaultOptionsContext, SuspenseDefaultOptionsContext } from './DefaultOptionsContexts'
+export { SuspensiveDevMode, useDevModeObserve, SuspensiveDevModeContext } from './SuspensiveDevModeContext'
