+++ packages/react/src/utility-types/PropsWithDevMode.ts
@@ -1,0 +1,6 @@
+export type PropsWithDevMode<TDevModeProps extends Record<string, any>> = {
+  /**
+   * @experimental This is experimental feature.
+   */
+  devMode?: TDevModeProps
+}
