+++ packages/react/src/utility-types/PropsWithoutDevMode.ts
@@ -1,0 +1,3 @@
+import type { ComponentProps, ComponentType } from 'react'
+
+export type PropsWithoutDevMode<TProps extends ComponentProps<ComponentType>> = Omit<TProps, 'devMode'>
