+++ packages/react/src/utility-types/index.ts
@@ -1,1 +1,3 @@
 export type { PropsWithoutChildren } from './PropsWithoutChildren'
+export type { PropsWithDevMode } from './PropsWithDevMode'
+export type { PropsWithoutDevMode } from './PropsWithoutDevMode'
